fix: context now own a copy of string parameters
Previously, the context would hold pointers to string parameters it does
own. This poses a risk of dangling pointers if the caller deallocated or
modified the memory while the context relied on it. To ensure safety,
usuability and avoid undefined behavior, the context now creates and
owns a copy a copy of all string parameters.

Changes:
  - Introduce an arena allocator.
  - Add two new error codes: DD_EALLOCATION and
    DD_ESTR_PARAM_EXCEED_MAX_LENGTH.
  - Enforce a limit on string parameters size of 1024 characters.